+++ src/layout/Container/Container.js
@@ -20,6 +20,10 @@ const Container = styled.div`
   ${p => paddingHorizontal(`calc(${p.theme.gridGutterWidth} / 2)`)};
   ${marginHorizontal('auto')};
   ${p => !p.fluid && responsiveCss};
+  &::before, &::after {
+    content: " ";
+    display: table;
+  }
 `;
 
 Container.propTypes = {